requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
instaloader==4.11
pyTelegramBotAPI==4.18.0
python-dotenv==1.0.1
requests==2.32.1
urllib3==2.2.1
